requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
distlib==0.3.8
Django==5.0.6
django-admin-tools==0.9.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
fonttools==4.53.1
geographiclib==2.0
geopy==2.4.1
html5lib==1.1
kombu==5.3.7
pillow==10.3.0
platformdirs==4.2.1
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.10.0
PyJWT==2.8.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.7
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
tzdata==2024.1
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
wheel==0.43.0
whitenoise==6.7.0
zopfli==0.2.3
django-cors-headers
